Donot defer and increment serial for same message
WifiController considers enable and disable messages as TOGGLE.
When WiFi is enabled immediately after disable, TOGGLE message is
deferred for about 500ms giving a minimum delay between TOGGLEs.
However if TOGGLE message is again sent to WifiController, it
defers the message and increments the serial to discard both deferred
TOGGLE messages thereby not enabling WiFi.

Second TOGGLE message should be checked if its the same enable message.
If it is, then donot defer the message and donot increment the serial.
This makes sure WiFi is enabled by the first deferred message.

Change-Id: I31fc9f3a5f084973495aaac951d08cbf6e48995a