Stop ANQP parsing when content of element is empty
If 802.11u/HS2.0 elements exist in ANQP response, framework tries to
parse content of element. When it's empty, Exception is sometimes
thrown and it results in that framework doesn't try to connect. This
issue happens at for instance IPAddressTypeAvailabilityElement when
the content of IP Address Type Availability ANQP-element is nothing.

This patch just skips parsing of the element if its content is empty.

Bug: 28336506
Change-Id: I9d5800b88965df1d736cababe4f82c674d7c65be